fix: merge misplaced settings test files into correct locations
- Merged LogsPage.branch.test.tsx into LogsPage.coverage.test.tsx (components/__tests__/)
- Merged ServiceConfig.branches.test.ts.tsx unique tests into ServiceConfig.test.tsx (components/__tests__/)
- Deleted the two misplaced files from settings/__tests__/
- Fixed Dialog mock to always render children so trigger buttons are accessible in tests
- All 312 settings tests passing
