--- frontend/src/features/contracts/__tests__/ArchiveTab.test.tsx
@@ -1,717 +1,0 @@
-import { render, screen, fireEvent, waitFor } from '@testing-library/react'
-import { ArchiveTab } from '../components/ArchiveTab'
-import { contractApi } from '../api'
-import type { Contract, ArchiveChecklist } from '../types'
-
-// ── Mocks ──
-
-vi.mock('lucide-react', () => ({
-  Upload: (p: Record<string, unknown>) => <svg data-testid="upload" {...p} />,
-  Trash2: (p: Record<string, unknown>) => <svg data-testid="trash" {...p} />,
-  Archive: (p: Record<string, unknown>) => <svg data-testid="archive" {...p} />,
-  FolderSync: (p: Record<string, unknown>) => <svg data-testid="folder-sync" {...p} />,
-  GripVertical: (p: Record<string, unknown>) => <svg data-testid="grip" {...p} />,
-  FileCheck: (p: Record<string, unknown>) => <svg data-testid="file-check" {...p} />,
-  Loader2: (p: Record<string, unknown>) => <svg data-testid="loader" {...p} />,
-  Scaling: (p: Record<string, unknown>) => <svg data-testid="scaling" {...p} />,
-  ArrowRightLeft: (p: Record<string, unknown>) => <svg data-testid="arrow-right-left" {...p} />,
-  ChevronDown: (p: Record<string, unknown>) => <svg data-testid="chevron-down" {...p} />,
-  ChevronRight: (p: Record<string, unknown>) => <svg data-testid="chevron-right" {...p} />,
-  Download: (p: Record<string, unknown>) => <svg data-testid="download" {...p} />,
-  Eye: (p: Record<string, unknown>) => <svg data-testid="eye" {...p} />,
-  FolderOpen: (p: Record<string, unknown>) => <svg data-testid="folder-open" {...p} />,
-  Sparkles: (p: Record<string, unknown>) => <svg data-testid="sparkles" {...p} />,
-}))
-
-vi.mock('sonner', () => ({
-  toast: { success: vi.fn(), error: vi.fn() },
-}))
-
-vi.mock('../api', () => ({
-  contractApi: {
-    getArchiveChecklist: vi.fn(),
-    uploadArchiveItem: vi.fn(),
-    deleteArchiveMaterial: vi.fn(),
-    syncCaseMaterials: vi.fn(),
-    confirmArchive: vi.fn(),
-    toggleCompactArchive: vi.fn(),
-    scaleToA4: vi.fn(),
-    reorderArchiveMaterials: vi.fn(),
-    moveArchiveMaterial: vi.fn(),
-    clearAllArchiveMaterials: vi.fn(),
-    generateArchiveFolder: vi.fn(),
-    learnArchiveRules: vi.fn(),
-    previewArchiveItem: vi.fn(),
-    downloadArchiveItem: vi.fn(),
-    previewSingleMaterial: vi.fn(),
-    previewArchivePlaceholders: vi.fn(),
-  },
-}))
-
-vi.mock('@dnd-kit/core', () => ({
-  DndContext: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  closestCorners: vi.fn(),
-  KeyboardSensor: vi.fn(),
-  PointerSensor: vi.fn(),
-  useSensor: vi.fn(),
-  useSensors: vi.fn(() => []),
-}))
-
-vi.mock('@dnd-kit/sortable', () => ({
-  SortableContext: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  useSortable: () => ({
-    attributes: {},
-    listeners: {},
-    setNodeRef: vi.fn(),
-    transform: null,
-    transition: null,
-    isDragging: false,
-  }),
-  verticalListSortingStrategy: vi.fn(),
-  arrayMove: (arr: unknown[], from: number, to: number) => {
-    const result = [...arr]
-    const [item] = result.splice(from, 1)
-    result.splice(to, 0, item)
-    return result
-  },
-  sortableKeyboardCoordinates: vi.fn(),
-}))
-
-vi.mock('@dnd-kit/utilities', () => ({
-  CSS: { Transform: { toString: () => null } },
-}))
-
-vi.mock('../components/FolderScanPanel', () => ({
-  FolderScanPanel: () => <div data-testid="folder-scan-panel" />,
-}))
-
-vi.mock('@/components/ui/button', () => ({
-  Button: ({ children, ...props }: Record<string, unknown>) => <button {...props}>{children}</button>,
-}))
-
-vi.mock('@/components/ui/select', () => ({
-  Select: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectItem: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectValue: () => <span />,
-}))
-
-vi.mock('@/components/ui/alert-dialog', () => ({
-  AlertDialog: ({ children, open }: { children: React.ReactNode; open?: boolean }) =>
-    open ? <div data-testid="alert-dialog">{children}</div> : null,
-  AlertDialogAction: ({ children, ...props }: Record<string, unknown>) => <button {...props}>{children}</button>,
-  AlertDialogCancel: ({ children, ...props }: Record<string, unknown>) => <button {...props}>{children}</button>,
-  AlertDialogContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogDescription: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogFooter: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogHeader: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogTitle: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-
-vi.mock('@/components/ui/dialog', () => ({
-  Dialog: ({ children, open }: { children: React.ReactNode; open?: boolean }) =>
-    open ? <div data-testid="dialog">{children}</div> : null,
-  DialogContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  DialogDescription: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  DialogHeader: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  DialogTitle: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-
-// ── Helpers ──
-
-function makeChecklist(overrides: Partial<ArchiveChecklist> = {}): ArchiveChecklist {
-  return {
-    archive_category: 'civil',
-    archive_category_label: '民事',
-    compact_archive: false,
-    items: [
-      {
-        code: 'contract_original',
-        name: '合同原件',
-        template: null,
-        required: true,
-        auto_detect: null,
-        source: 'manual',
-        completed: true,
-        material_ids: [1],
-        materials: [
-          { id: 1, original_filename: 'contract.pdf', source: 'upload', source_label: '', category: 'contract_original', category_label: '', filename: 'contract.pdf', file_url: '', file_size: 1024, order: 0, archive_item_code: 'contract_original', remark: null, uploaded_at: null, created_at: null },
-        ],
-        has_case_material: false,
-      },
-      {
-        code: 'auth_doc',
-        name: '授权材料',
-        template: null,
-        required: false,
-        auto_detect: 'supervision_card',
-        source: 'manual',
-        completed: false,
-        material_ids: [],
-        materials: [],
-        has_case_material: true,
-      },
-      {
-        code: 'template_doc',
-        name: '模板文书',
-        template: 'template_sub',
-        required: false,
-        auto_detect: null,
-        source: 'manual',
-        completed: false,
-        material_ids: [],
-        materials: [],
-        has_case_material: false,
-      },
-    ],
-    completed_count: 1,
-    total_count: 3,
-    required_completed_count: 1,
-    required_total_count: 1,
-    completion_percentage: 100,
-    ...overrides,
-  }
-}
-
-function makeContract(overrides: Partial<Contract> = {}): Contract {
-  return {
-    id: 1,
-    name: 'Test Contract',
-    status: 'active',
-    ...overrides,
-  } as unknown as Contract
-}
-
-describe('ArchiveTab', () => {
-  beforeEach(() => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist())
-    vi.mocked(contractApi.syncCaseMaterials).mockResolvedValue({ synced_count: 3 } as never)
-    vi.mocked(contractApi.confirmArchive).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.toggleCompactArchive).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.scaleToA4).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.clearAllArchiveMaterials).mockResolvedValue({ deleted_count: 5 } as never)
-    vi.mocked(contractApi.generateArchiveFolder).mockResolvedValue({ success: true, generated_docs: ['doc1'], errors: [] } as never)
-    vi.mocked(contractApi.learnArchiveRules).mockResolvedValue({ success: true, message: '学习完成' } as never)
-    vi.mocked(contractApi.deleteArchiveMaterial).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.moveArchiveMaterial).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.uploadArchiveItem).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.reorderArchiveMaterials).mockResolvedValue(undefined as never)
-    vi.mocked(contractApi.previewArchivePlaceholders).mockResolvedValue({ success: true, data: [{ key: 'K1', label: 'Label1', value: 'Val1' }] } as never)
-  })
-
-  afterEach(() => {
-    vi.restoreAllMocks()
-  })
-
-  it('renders checklist title', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('归档检查清单')).toBeInTheDocument())
-  })
-
-  it('fetches checklist on mount', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(contractApi.getArchiveChecklist).toHaveBeenCalledWith(1))
-  })
-
-  it('renders checklist items', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('合同原件').length).toBeGreaterThan(0))
-    expect(screen.getAllByText('授权材料').length).toBeGreaterThan(0)
-    expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0)
-  })
-
-  it('renders progress percentage', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText(/1\/2/).length).toBeGreaterThan(0))
-  })
-
-  it('shows required count when not all required done', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({
-      required_completed_count: 0,
-      required_total_count: 1,
-    }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText(/必需项: 0\/1/)).toBeInTheDocument())
-  })
-
-  it('renders sync case materials button', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('从案件材料同步')).toBeInTheDocument())
-  })
-
-  it('renders scale to A4 button', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('缩放至A4')).toBeInTheDocument())
-  })
-
-  it('renders generate folder button', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('生成归档文件夹')).toBeInTheDocument())
-  })
-
-  it('renders learn rules button', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('学习分类规则')).toBeInTheDocument())
-  })
-
-  it('renders folder scan button', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('从合同文件夹同步')).toBeInTheDocument())
-  })
-
-  it('shows confirm archive button when can archive', async () => {
-    render(<ArchiveTab contract={makeContract({ status: 'active' })} />)
-    await waitFor(() => expect(screen.getByText('确认归档')).toBeInTheDocument())
-  })
-
-  it('hides confirm archive when status is archived', async () => {
-    render(<ArchiveTab contract={makeContract({ status: 'archived' })} />)
-    await waitFor(() => expect(contractApi.getArchiveChecklist).toHaveBeenCalled())
-    expect(screen.queryByText('确认归档')).not.toBeInTheDocument()
-  })
-
-  it('handles sync case materials', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('从案件材料同步')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('从案件材料同步'))
-    await waitFor(() => expect(contractApi.syncCaseMaterials).toHaveBeenCalledWith(1))
-  })
-
-  it('handles sync case materials error', async () => {
-    vi.mocked(contractApi.syncCaseMaterials).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('从案件材料同步')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('从案件材料同步'))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('同步失败'))
-  })
-
-  it('handles scale to A4', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('缩放至A4')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('缩放至A4'))
-    await waitFor(() => expect(contractApi.scaleToA4).toHaveBeenCalledWith(1))
-  })
-
-  it('handles scale to A4 error', async () => {
-    vi.mocked(contractApi.scaleToA4).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('缩放至A4')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('操作失败'))
-  })
-
-  it('handles generate folder success with docs', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('生成归档文件夹')))
-    await waitFor(() => expect(contractApi.generateArchiveFolder).toHaveBeenCalledWith(1))
-  })
-
-  it('handles generate folder success with no docs', async () => {
-    vi.mocked(contractApi.generateArchiveFolder).mockResolvedValue({ success: true, generated_docs: [], errors: [] } as never)
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('生成归档文件夹')))
-    await waitFor(() => expect(contractApi.generateArchiveFolder).toHaveBeenCalled())
-  })
-
-  it('handles generate folder failure', async () => {
-    vi.mocked(contractApi.generateArchiveFolder).mockResolvedValue({ success: false, generated_docs: [], errors: ['Error msg'] } as never)
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('生成归档文件夹')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('Error msg'))
-  })
-
-  it('handles generate folder failure with no error message', async () => {
-    vi.mocked(contractApi.generateArchiveFolder).mockResolvedValue({ success: false, generated_docs: [], errors: [] } as never)
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('生成归档文件夹')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('生成失败'))
-  })
-
-  it('handles generate folder exception', async () => {
-    vi.mocked(contractApi.generateArchiveFolder).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('生成归档文件夹')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('生成归档文件夹失败'))
-  })
-
-  it('handles learn rules success', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('学习分类规则')))
-    await waitFor(() => expect(contractApi.learnArchiveRules).toHaveBeenCalled())
-  })
-
-  it('handles learn rules failure result', async () => {
-    vi.mocked(contractApi.learnArchiveRules).mockResolvedValue({ success: false, message: '学习失败' } as never)
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('学习分类规则')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('学习失败'))
-  })
-
-  it('handles learn rules failure with empty message', async () => {
-    vi.mocked(contractApi.learnArchiveRules).mockResolvedValue({ success: false, message: '' } as never)
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('学习分类规则')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('学习失败'))
-  })
-
-  it('handles learn rules exception', async () => {
-    vi.mocked(contractApi.learnArchiveRules).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('学习分类规则')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('学习分类规则失败'))
-  })
-
-  it('handles toggle compact', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('归档检查清单')).toBeInTheDocument())
-    // The compact toggle is a button with a title
-    const compactBtn = screen.getByTitle('精简视图')
-    fireEvent.click(compactBtn)
-    await waitFor(() => expect(contractApi.toggleCompactArchive).toHaveBeenCalledWith(1))
-  })
-
-  it('handles toggle compact error', async () => {
-    vi.mocked(contractApi.toggleCompactArchive).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByTitle('精简视图')))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('操作失败'))
-  })
-
-  it('renders compact mode badge when compact_archive is true', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({ compact_archive: true }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByTitle('显示全部')).toBeInTheDocument())
-  })
-
-  it('opens folder scan dialog', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => fireEvent.click(screen.getByText('从合同文件夹同步')))
-    expect(screen.getByTestId('folder-scan-panel')).toBeInTheDocument()
-  })
-
-  it('renders template items with preview and download buttons', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0))
-  })
-
-  it('renders completed items with preview and download buttons', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('合同原件').length).toBeGreaterThan(0))
-  })
-
-  it('renders material source labels', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({
-      items: [{
-        code: 'test',
-        name: 'Test Item',
-        template: null,
-        required: false,
-        auto_detect: null,
-        source: 'manual',
-        completed: true,
-        material_ids: [1],
-        materials: [
-          { id: 1, original_filename: 'file.pdf', source: 'case', source_label: '案件材料', category: 'other', category_label: '', filename: 'file.pdf', file_url: '', file_size: 1024, order: 0, archive_item_code: 'test', remark: null, uploaded_at: null, created_at: null },
-          { id: 2, original_filename: 'scan.pdf', source: 'scan', source_label: '扫描', category: 'other', category_label: '', filename: 'scan.pdf', file_url: '', file_size: 512, order: 1, archive_item_code: 'test', remark: null, uploaded_at: null, created_at: null },
-          { id: 3, original_filename: 'other.pdf', source: 'other', source_label: null, category: 'other', category_label: '', filename: 'other.pdf', file_url: '', file_size: 256, order: 2, archive_item_code: 'test', remark: null, uploaded_at: null, created_at: null },
-        ],
-        has_case_material: false,
-      }],
-    }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('案件材料')).toBeInTheDocument())
-    expect(screen.getByText('扫描')).toBeInTheDocument()
-  })
-
-  it('renders ItemBadge for completed item', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('合同原件')).toBeInTheDocument())
-    // The badge should show a checkmark for completed items
-  })
-
-  it('renders ItemBadge for template item', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0))
-  })
-
-  it('renders ItemBadge for auto_detect item', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('授权材料').length).toBeGreaterThan(0))
-  })
-
-  it('renders ItemBadge for has_case_material item', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('授权材料').length).toBeGreaterThan(0))
-  })
-
-  it('renders ItemBadge for required but not completed item', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({
-      items: [{
-        code: 'req',
-        name: 'Required Item',
-        template: null,
-        required: true,
-        auto_detect: null,
-        source: 'manual',
-        completed: false,
-        material_ids: [],
-        materials: [],
-        has_case_material: false,
-      }],
-    }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('Required Item')).toBeInTheDocument())
-  })
-
-  it('renders ItemBadge for optional non-special item', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({
-      items: [{
-        code: 'opt',
-        name: 'Optional Item',
-        template: null,
-        required: false,
-        auto_detect: null,
-        source: 'manual',
-        completed: false,
-        material_ids: [],
-        materials: [],
-        has_case_material: false,
-      }],
-    }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('Optional Item')).toBeInTheDocument())
-  })
-
-  it('handles compact mode filter correctly', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({
-      compact_archive: true,
-      items: [
-        { code: 'template', name: 'Template', template: 'sub', required: false, auto_detect: null, source: 'manual', completed: false, material_ids: [], materials: [], has_case_material: false },
-        { code: 'completed', name: 'Completed', template: null, required: false, auto_detect: null, source: 'manual', completed: true, material_ids: [], materials: [], has_case_material: false },
-        { code: 'required', name: 'Required', template: null, required: true, auto_detect: null, source: 'manual', completed: false, material_ids: [], materials: [], has_case_material: false },
-        { code: 'optional_empty', name: 'Optional Empty', template: null, required: false, auto_detect: null, source: 'manual', completed: false, material_ids: [], materials: [], has_case_material: false },
-      ],
-    }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('Template')).toBeInTheDocument())
-    expect(screen.getByText('Completed')).toBeInTheDocument()
-    expect(screen.getByText('Required')).toBeInTheDocument()
-    expect(screen.queryByText('Optional Empty')).not.toBeInTheDocument()
-  })
-
-  it('handles fetchChecklist error', async () => {
-    vi.mocked(contractApi.getArchiveChecklist).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('获取检查清单失败'))
-  })
-
-  it('opens confirm archive dialog', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('确认归档')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('确认归档'))
-    await waitFor(() => expect(screen.getByText('确认归档后，合同状态将变为「已归档」，关联的案件将自动关闭。此操作不可逆。')).toBeInTheDocument())
-  })
-
-  it('handles confirm archive', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('确认归档')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('确认归档'))
-    await waitFor(() => expect(screen.getAllByText('确认归档').length).toBeGreaterThan(1))
-    const buttons = screen.getAllByText('确认归档')
-    fireEvent.click(buttons[buttons.length - 1])
-    await waitFor(() => expect(contractApi.confirmArchive).toHaveBeenCalledWith(1))
-  })
-
-  it('handles confirm archive error', async () => {
-    vi.mocked(contractApi.confirmArchive).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('确认归档')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('确认归档'))
-    await waitFor(() => expect(screen.getAllByText('确认归档').length).toBeGreaterThan(1))
-    const buttons = screen.getAllByText('确认归档')
-    fireEvent.click(buttons[buttons.length - 1])
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('归档确认失败'))
-  })
-
-  it('opens clear all dialog', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByTitle('清空全部材料')).toBeInTheDocument())
-    fireEvent.click(screen.getByTitle('清空全部材料'))
-    await waitFor(() => expect(screen.getByText('确认清空全部材料')).toBeInTheDocument())
-  })
-
-  it('handles clear all', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByTitle('清空全部材料')).toBeInTheDocument())
-    fireEvent.click(screen.getByTitle('清空全部材料'))
-    await waitFor(() => expect(screen.getByText('清空全部')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('清空全部'))
-    await waitFor(() => expect(contractApi.clearAllArchiveMaterials).toHaveBeenCalledWith(1))
-  })
-
-  it('handles clear all error', async () => {
-    vi.mocked(contractApi.clearAllArchiveMaterials).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByTitle('清空全部材料')).toBeInTheDocument())
-    fireEvent.click(screen.getByTitle('清空全部材料'))
-    await waitFor(() => expect(screen.getByText('清空全部')).toBeInTheDocument())
-    fireEvent.click(screen.getByText('清空全部'))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('清空失败'))
-  })
-
-  it('opens delete material dialog when delete icon clicked', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('合同原件').length).toBeGreaterThan(0))
-    const deleteButtons = screen.getAllByTitle('删除')
-    if (deleteButtons.length > 0) {
-      fireEvent.click(deleteButtons[0])
-      await waitFor(() => expect(screen.getByText('确认删除材料')).toBeInTheDocument())
-    }
-  })
-
-  it('handles delete material', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('合同原件').length).toBeGreaterThan(0))
-    const deleteButtons = screen.getAllByTitle('删除')
-    if (deleteButtons.length > 0) {
-      fireEvent.click(deleteButtons[0])
-      await waitFor(() => expect(screen.getByText('确认删除材料')).toBeInTheDocument())
-    }
-  })
-
-  it('handles preview archive item click', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('合同原件').length).toBeGreaterThan(0))
-    const previewButtons = screen.getAllByTitle('预览')
-    if (previewButtons.length > 0) {
-      fireEvent.click(previewButtons[0])
-      expect(contractApi.previewArchiveItem).toHaveBeenCalledWith(1, 'contract_original')
-    }
-  })
-
-  it('handles download archive item click', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('合同原件').length).toBeGreaterThan(0))
-    const downloadButtons = screen.getAllByTitle('下载材料')
-    if (downloadButtons.length > 0) {
-      fireEvent.click(downloadButtons[0])
-      expect(contractApi.downloadArchiveItem).toHaveBeenCalled()
-    }
-  })
-
-  it('handles upload button click', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('授权材料').length).toBeGreaterThan(0))
-    const uploadButtons = screen.getAllByTitle('上传文件')
-    expect(uploadButtons.length).toBeGreaterThan(0)
-  })
-
-  it('handles placeholder preview for template items', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0))
-    const previewButtons = screen.getAllByTitle('预览替换词')
-    if (previewButtons.length > 0) {
-      fireEvent.click(previewButtons[0])
-      await waitFor(() => expect(contractApi.previewArchivePlaceholders).toHaveBeenCalledWith(1, 'template_sub'))
-    }
-  })
-
-  it('handles placeholder preview failure', async () => {
-    vi.mocked(contractApi.previewArchivePlaceholders).mockResolvedValue({ success: false, error: 'Preview failed' } as never)
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0))
-    const previewButtons = screen.getAllByTitle('预览替换词')
-    if (previewButtons.length > 0) {
-      fireEvent.click(previewButtons[0])
-      await waitFor(() => expect(toast.error).toHaveBeenCalledWith('Preview failed'))
-    }
-  })
-
-  it('handles placeholder preview with empty error', async () => {
-    vi.mocked(contractApi.previewArchivePlaceholders).mockResolvedValue({ success: false, error: '' } as never)
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0))
-    const previewButtons = screen.getAllByTitle('预览替换词')
-    if (previewButtons.length > 0) {
-      fireEvent.click(previewButtons[0])
-      await waitFor(() => expect(toast.error).toHaveBeenCalledWith('预览失败'))
-    }
-  })
-
-  it('handles placeholder preview exception', async () => {
-    vi.mocked(contractApi.previewArchivePlaceholders).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('模板文书').length).toBeGreaterThan(0))
-    const previewButtons = screen.getAllByTitle('预览替换词')
-    if (previewButtons.length > 0) {
-      fireEvent.click(previewButtons[0])
-      await waitFor(() => expect(toast.error).toHaveBeenCalledWith('预览请求失败'))
-    }
-  })
-
-  it('handles getMaterialsForCode with missing code', async () => {
-    // Tests the fallback when item code is not found
-    vi.mocked(contractApi.getArchiveChecklist).mockResolvedValue(makeChecklist({
-      items: [{
-        code: 'no_materials',
-        name: 'No Materials',
-        template: null,
-        required: false,
-        auto_detect: null,
-        source: 'manual',
-        completed: false,
-        material_ids: [],
-        materials: [],
-        has_case_material: false,
-      }],
-    }))
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('No Materials')).toBeInTheDocument())
-  })
-
-  it('renders non-template items count correctly', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('归档检查清单')).toBeInTheDocument())
-    // non-template items = 2 (contract_original + auth_doc), template = 1
-    const countBadges = screen.getAllByText(/1\/2/)
-    expect(countBadges.length).toBeGreaterThan(0)
-  })
-
-  it('handles expand all toggle', async () => {
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('归档检查清单')).toBeInTheDocument())
-    const expandBtn = screen.getByTitle('展开全部子项')
-    fireEvent.click(expandBtn)
-    // Should toggle expand state
-  })
-
-  it('handles move material', async () => {
-    // The move action is handled via Select onValueChange in SortableMaterialItem
-    // We test the handleMoveMaterial callback indirectly
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getByText('合同原件')).toBeInTheDocument())
-  })
-
-  it('handles upload error', async () => {
-    vi.mocked(contractApi.uploadArchiveItem).mockRejectedValue(new Error('fail'))
-    const { toast } = await import('sonner')
-    render(<ArchiveTab contract={makeContract()} />)
-    await waitFor(() => expect(screen.getAllByText('授权材料').length).toBeGreaterThan(0))
-    // Upload is triggered via file input which is hard to simulate directly
-  })
-})
