--- frontend/src/features/contracts/__tests__/ContractForm.test.tsx
@@ -1,291 +1,0 @@
-import { render, screen, fireEvent, waitFor } from '@testing-library/react'
-import { ContractForm } from '../components/ContractForm'
-import type { Contract } from '../types'
-
-// ── Mocks ──
-
-vi.mock('lucide-react', () => ({
-  Check: (p: Record<string, unknown>) => <svg data-testid="check" {...p} />,
-  ChevronsUpDown: (p: Record<string, unknown>) => <svg data-testid="chevron" {...p} />,
-}))
-
-vi.mock('sonner', () => ({
-  toast: { success: vi.fn(), error: vi.fn() },
-}))
-
-vi.mock('react-router', () => ({
-  useNavigate: () => vi.fn(),
-}))
-
-vi.mock('@/routes/paths', () => ({
-  PATHS: { ADMIN_CONTRACTS: '/admin/contracts' },
-}))
-
-vi.mock('../hooks/use-contract-mutations', () => ({
-  useContractMutations: () => ({
-    createContract: { mutateAsync: vi.fn().mockResolvedValue({ id: 1 }) },
-    updateContract: { mutateAsync: vi.fn().mockResolvedValue({ id: 1 }) },
-  }),
-}))
-
-vi.mock('../hooks/use-lawyers', () => ({
-  useLawyers: () => ({
-    data: [
-      { id: 1, username: 'lawyer1', real_name: '张律师' },
-      { id: 2, username: 'lawyer2', real_name: '李律师' },
-    ],
-  }),
-}))
-
-vi.mock('../hooks/use-clients-select', () => ({
-  useClientsSelect: () => ({
-    data: [
-      { id: 1, name: 'Client A' },
-      { id: 2, name: 'Client B' },
-    ],
-  }),
-}))
-
-vi.mock('@/lib/utils', () => ({
-  cn: (...args: unknown[]) => args.filter(Boolean).join(' '),
-}))
-
-vi.mock('@/components/ui/button', () => ({
-  Button: ({ children, ...props }: Record<string, unknown>) => <button {...props}>{children}</button>,
-}))
-
-vi.mock('@/components/ui/input', () => ({
-  Input: (props: Record<string, unknown>) => <input {...props} />,
-}))
-
-vi.mock('@/components/ui/label', () => ({
-  Label: ({ children, ...props }: Record<string, unknown>) => <label {...props}>{children}</label>,
-}))
-
-vi.mock('@/components/ui/card', () => ({
-  Card: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-  CardContent: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-  CardHeader: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-  CardTitle: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-}))
-
-vi.mock('@/components/ui/select', () => ({
-  Select: ({ children, onValueChange, value }: { children: React.ReactNode; onValueChange?: (v: string) => void; value?: string }) => (
-    <div data-testid="select" data-value={value}>{children}</div>
-  ),
-  SelectContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectItem: ({ children, value }: { children: React.ReactNode; value: string }) => (
-    <div data-testid={`select-item-${value}`}>{children}</div>
-  ),
-  SelectTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectValue: () => <span />,
-}))
-
-vi.mock('@/components/ui/badge', () => ({
-  Badge: ({ children, ...props }: Record<string, unknown>) => <span {...props}>{children}</span>,
-}))
-
-vi.mock('@/components/ui/command', () => ({
-  Command: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  CommandEmpty: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  CommandGroup: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  CommandInput: (props: Record<string, unknown>) => <input {...props} />,
-  CommandItem: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-  CommandList: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-
-vi.mock('@/components/ui/popover', () => ({
-  Popover: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  PopoverContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  PopoverTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-
-describe('ContractForm', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-  })
-
-  it('renders create mode form', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('基本信息')).toBeInTheDocument()
-    expect(screen.getByText('创建合同')).toBeInTheDocument()
-  })
-
-  it('renders edit mode form with contract data', () => {
-    const contract = {
-      id: 1,
-      name: 'Existing Contract',
-      case_type: 'civil',
-      fee_mode: 'FIXED',
-      fixed_amount: 10000,
-      risk_rate: 10,
-      custom_terms: null,
-      specified_date: '2024-01-01',
-      start_date: '2024-01-01',
-      end_date: '2024-12-31',
-      assignments: [{ lawyer_id: 1 }],
-      contract_parties: [{ client: 1, role: 'PRINCIPAL' }],
-    } as unknown as Contract
-    render(<ContractForm mode="edit" contract={contract} />)
-    expect(screen.getByText('保存修改')).toBeInTheDocument()
-  })
-
-  it('shows error when name is empty', async () => {
-    const { toast } = await import('sonner')
-    render(<ContractForm mode="create" />)
-    fireEvent.click(screen.getByText('创建合同'))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('请输入合同名称'))
-  })
-
-  it('shows error when no lawyers selected', async () => {
-    const { toast } = await import('sonner')
-    render(<ContractForm mode="create" />)
-    // Fill in name
-    const nameInput = screen.getByPlaceholderText('输入合同名称')
-    fireEvent.change(nameInput, { target: { value: 'Test' } })
-    fireEvent.click(screen.getByText('创建合同'))
-    await waitFor(() => expect(toast.error).toHaveBeenCalledWith('请至少指派一个律师'))
-  })
-
-  it('renders lawyer badges', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('张律师')).toBeInTheDocument()
-    expect(screen.getByText('李律师')).toBeInTheDocument()
-  })
-
-  it('toggles lawyer selection', () => {
-    render(<ContractForm mode="create" />)
-    fireEvent.click(screen.getByText('张律师'))
-    // Should toggle
-  })
-
-  it('renders add party button', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('添加')).toBeInTheDocument()
-  })
-
-  it('adds a party on click', () => {
-    render(<ContractForm mode="create" />)
-    fireEvent.click(screen.getByText('添加'))
-    expect(screen.getByText('选择当事人')).toBeInTheDocument()
-  })
-
-  it('removes a party', () => {
-    render(<ContractForm mode="create" />)
-    fireEvent.click(screen.getByText('添加'))
-    fireEvent.click(screen.getByText('×'))
-    expect(screen.queryByText('选择当事人')).not.toBeInTheDocument()
-  })
-
-  it('renders cancel button', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('取消')).toBeInTheDocument()
-  })
-
-  it('shows empty parties message', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('未添加当事人')).toBeInTheDocument()
-  })
-
-  it('renders fee mode select', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('收费信息')).toBeInTheDocument()
-  })
-
-  it('shows fixed amount for FIXED mode', () => {
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('固定金额')).toBeInTheDocument()
-  })
-
-  it('shows risk rate for SEMI_RISK mode', () => {
-    const contract = {
-      id: 1,
-      name: 'Test',
-      case_type: 'civil',
-      fee_mode: 'SEMI_RISK',
-      fixed_amount: null,
-      risk_rate: null,
-      custom_terms: null,
-      specified_date: null,
-      start_date: null,
-      end_date: null,
-      assignments: [{ lawyer_id: 1 }],
-      contract_parties: [],
-    } as unknown as Contract
-    render(<ContractForm mode="edit" contract={contract} />)
-    expect(screen.getByText('前期金额')).toBeInTheDocument()
-    expect(screen.getByText('风险比例(%)')).toBeInTheDocument()
-  })
-
-  it('shows custom terms for CUSTOM mode', () => {
-    const contract = {
-      id: 1,
-      name: 'Test',
-      case_type: 'civil',
-      fee_mode: 'CUSTOM',
-      fixed_amount: null,
-      risk_rate: null,
-      custom_terms: 'Some terms',
-      specified_date: null,
-      start_date: null,
-      end_date: null,
-      assignments: [{ lawyer_id: 1 }],
-      contract_parties: [],
-    } as unknown as Contract
-    render(<ContractForm mode="edit" contract={contract} />)
-    expect(screen.getByText('自定义条款')).toBeInTheDocument()
-  })
-
-  it('shows risk rate for FULL_RISK mode', () => {
-    const contract = {
-      id: 1,
-      name: 'Test',
-      case_type: 'civil',
-      fee_mode: 'FULL_RISK',
-      fixed_amount: null,
-      risk_rate: null,
-      custom_terms: null,
-      specified_date: null,
-      start_date: null,
-      end_date: null,
-      assignments: [{ lawyer_id: 1 }],
-      contract_parties: [],
-    } as unknown as Contract
-    render(<ContractForm mode="edit" contract={contract} />)
-    expect(screen.getByText('风险比例(%)')).toBeInTheDocument()
-  })
-
-  it('shows lawyer with real_name fallback to username', () => {
-    // The mock returns real_name='张律师' which is used by the component
-    render(<ContractForm mode="create" />)
-    expect(screen.getByText('张律师')).toBeInTheDocument()
-  })
-
-  it('renders party with selected client name', () => {
-    const contract = {
-      id: 1,
-      name: 'Test',
-      case_type: 'civil',
-      fee_mode: 'FIXED',
-      fixed_amount: null,
-      risk_rate: null,
-      custom_terms: null,
-      specified_date: null,
-      start_date: null,
-      end_date: null,
-      assignments: [{ lawyer_id: 1 }],
-      contract_parties: [{ client: 1, role: 'PRINCIPAL' }],
-    } as unknown as Contract
-    render(<ContractForm mode="edit" contract={contract} />)
-    const clientElements = screen.getAllByText('Client A')
-    expect(clientElements.length).toBeGreaterThan(0)
-  })
-
-  it('renders primary lawyer indicator', () => {
-    render(<ContractForm mode="create" />)
-    // The lawyer badges render real_name with (主办) for the first selected
-    // Since no lawyers are selected by default in create mode, (主办) won't show
-    // But we can verify the lawyer section renders
-    expect(screen.getByText('律师指派 *')).toBeInTheDocument()
-  })
-})
