+++ frontend/src/features/contracts/components/__tests__/ContractInfoCard.test.tsx
