+++ frontend/src/features/contracts/components/__tests__/FolderScanPanel.test.tsx
