--- frontend/src/features/settings/__tests__/LogsPage.branch.test.tsx
@@ -1,321 +1,0 @@
-/**
- * Branch-focused tests for LogsPage.tsx
- * Targets uncovered branches in relativeDate, handleAdd, dialog form, etc.
- */
-import { render, screen, fireEvent } from '@testing-library/react'
-import { LogsPage } from '../components/LogsPage'
-
-vi.mock('lucide-react', () => ({
-  Search: (p: Record<string, unknown>) => <svg data-testid="search-icon" {...p} />,
-  Loader2: (p: Record<string, unknown>) => <svg data-testid="loader-icon" {...p} />,
-  Plus: (p: Record<string, unknown>) => <svg data-testid="plus-icon" {...p} />,
-  Clock: (p: Record<string, unknown>) => <svg data-testid="clock-icon" {...p} />,
-  Paperclip: (p: Record<string, unknown>) => <svg data-testid="paperclip" {...p} />,
-  Bell: (p: Record<string, unknown>) => <svg data-testid="bell-icon" {...p} />,
-  Trash2: (p: Record<string, unknown>) => <svg data-testid="trash-icon" {...p} />,
-}))
-
-vi.mock('@/lib/date', () => ({ formatDate: (d: string) => d ?? '' }))
-
-const mockCreateLog = vi.fn()
-const mockDeleteLog = vi.fn()
-const mockListAllLogs = vi.fn().mockResolvedValue([])
-const mockListCases = vi.fn().mockResolvedValue([])
-
-vi.mock('@/features/cases/api', () => ({
-  caseApi: {
-    listAllLogs: (...args: unknown[]) => mockListAllLogs(...args),
-    list: (...args: unknown[]) => mockListCases(...args),
-    createLog: (...args: unknown[]) => mockCreateLog(...args),
-    deleteLog: (...args: unknown[]) => mockDeleteLog(...args),
-  },
-}))
-
-vi.mock('@/features/cases/types', () => ({
-  CASE_LOG_REMINDER_TYPE_LABELS: {
-    hearing: { zh: '开庭' },
-    asset_preservation: { zh: '财产保全' },
-    evidence_deadline: { zh: '举证期限' },
-    other: { zh: '其他' },
-  },
-}))
-
-let mockQueryData: unknown[] = []
-let mockQueryLoading = false
-let mockMutationFns: Record<string, (...args: unknown[]) => void> = {}
-
-vi.mock('@tanstack/react-query', async () => {
-  const actual = await vi.importActual<typeof import('@tanstack/react-query')>('@tanstack/react-query')
-  return {
-    ...actual,
-    useQuery: vi.fn(({ queryKey }: { queryKey: string[] }) => {
-      if (queryKey[0] === 'all-logs') {
-        return { data: mockQueryData, isLoading: mockQueryLoading }
-      }
-      return { data: [], isLoading: false }
-    }),
-    useMutation: vi.fn((config: { mutationFn: (...args: unknown[]) => unknown; onSuccess?: () => void }) => {
-      const mutate = vi.fn((...args: unknown[]) => {
-        mockMutationFns['mutate']?.(...args)
-        config.onSuccess?.()
-      })
-      return { mutate, isPending: false }
-    }),
-    useQueryClient: vi.fn(() => ({ invalidateQueries: vi.fn() })),
-    keepPreviousData: 'keepPreviousData',
-  }
-})
-
-vi.mock('@/components/ui/input', () => ({ Input: (props: Record<string, unknown>) => <input {...props} /> }))
-vi.mock('@/components/ui/label', () => ({ Label: ({ children, ...props }: Record<string, unknown>) => <label {...props}>{children}</label> }))
-vi.mock('@/components/ui/badge', () => ({ Badge: ({ children, ...props }: Record<string, unknown>) => <span {...props}>{children}</span> }))
-vi.mock('@/components/ui/button', () => ({ Button: ({ children, ...props }: Record<string, unknown>) => <button {...props}>{children}</button> }))
-vi.mock('@/components/ui/card', () => ({
-  Card: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-  CardContent: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-  CardHeader: ({ children, ...props }: Record<string, unknown>) => <div {...props}>{children}</div>,
-}))
-vi.mock('@/components/ui/select', () => ({
-  Select: ({ children, onValueChange, value }: { children: React.ReactNode; onValueChange?: (v: string) => void; value?: string }) => (
-    <div data-testid="select" data-value={value}>{children}</div>
-  ),
-  SelectContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectItem: ({ children, value }: { children: React.ReactNode; value: string }) => (
-    <button data-testid="select-item" data-value={value}>{children}</button>
-  ),
-  SelectTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  SelectValue: () => <span />,
-}))
-vi.mock('@/components/ui/dialog', () => ({
-  Dialog: ({ children, open }: { children: React.ReactNode; open?: boolean }) => <div data-testid="dialog" data-open={open}>{children}</div>,
-  DialogContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  DialogHeader: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  DialogTitle: ({ children }: { children: React.ReactNode }) => <h2>{children}</h2>,
-  DialogFooter: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  DialogTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-vi.mock('@/components/ui/alert-dialog', () => ({
-  AlertDialog: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogAction: ({ children, ...props }: Record<string, unknown>) => <button {...props}>{children}</button>,
-  AlertDialogCancel: ({ children }: { children: React.ReactNode }) => <button>{children}</button>,
-  AlertDialogContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogDescription: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogFooter: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogHeader: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogTitle: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  AlertDialogTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-vi.mock('@/components/ui/tooltip', () => ({
-  Tooltip: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  TooltipContent: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-  TooltipTrigger: ({ children }: { children: React.ReactNode }) => <div>{children}</div>,
-}))
-
-const makeLog = (overrides = {}) => ({
-  id: 1, case: 101, content: 'Test log', actor: 1,
-  actor_detail: { real_name: 'Zhang', username: 'zhang' },
-  attachments: [], reminders: [],
-  created_at: '2025-06-01 10:00:00', updated_at: '2025-06-01 10:00:00',
-  ...overrides,
-})
-
-describe('LogsPage - branch coverage', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mockQueryData = []
-    mockQueryLoading = false
-    mockMutationFns = {}
-  })
-
-  // relativeDate: days === 0 (branch 0[0])
-  it('renders today date for logs created today', () => {
-    const now = new Date()
-    const today = `${now.getFullYear()}-${String(now.getMonth() + 1).padStart(2, '0')}-${String(now.getDate()).padStart(2, '0')} 10:00:00`
-    mockQueryData = [makeLog({ created_at: today })]
-    render(<LogsPage />)
-    expect(screen.getByText('Test log')).toBeInTheDocument()
-  })
-
-  // relativeDate: days === 1 (branch 1[0])
-  it('renders yesterday for logs created yesterday', () => {
-    const yesterday = new Date(Date.now() - 86400000)
-    const dateStr = `${yesterday.getFullYear()}-${String(yesterday.getMonth() + 1).padStart(2, '0')}-${String(yesterday.getDate()).padStart(2, '0')} 10:00:00`
-    mockQueryData = [makeLog({ created_at: dateStr })]
-    render(<LogsPage />)
-    expect(screen.getByText('Test log')).toBeInTheDocument()
-  })
-
-  // relativeDate: days < 7 (branch 2[0])
-  it('renders days ago for recent logs', () => {
-    const recent = new Date(Date.now() - 3 * 86400000)
-    const dateStr = `${recent.getFullYear()}-${String(recent.getMonth() + 1).padStart(2, '0')}-${String(recent.getDate()).padStart(2, '0')} 10:00:00`
-    mockQueryData = [makeLog({ created_at: dateStr })]
-    render(<LogsPage />)
-    expect(screen.getByText('Test log')).toBeInTheDocument()
-  })
-
-  // relativeDate: days >= 7 (fallback: return dt.slice(0, 10))
-  it('renders date string for old logs', () => {
-    mockQueryData = [makeLog({ created_at: '2020-01-15 10:00:00' })]
-    render(<LogsPage />)
-    expect(screen.getByText('2020-01-15')).toBeInTheDocument()
-  })
-
-  // relativeDate: returns same as dateKey -> shows only dateKey (line 255-257)
-  it('shows relative date when different from dateKey', () => {
-    mockQueryData = [makeLog({ created_at: '2020-01-15 10:00:00' })]
-    render(<LogsPage />)
-    // When relativeDate returns the date string itself, the display is just the date
-    expect(screen.getByText('2020-01-15')).toBeInTheDocument()
-  })
-
-  // handleAdd: guard - no caseId or no content (branch 6: !newCaseId || !newContent.trim())
-  it('does not submit when caseId or content is empty', () => {
-    render(<LogsPage />)
-    const confirmBtn = screen.getByText('确认')
-    fireEvent.click(confirmBtn)
-    expect(mockCreateLog).not.toHaveBeenCalled()
-  })
-
-  // handleAdd: hasReminder truthy (branch 8: reminderType && reminderTime)
-  it('renders reminder settings section', () => {
-    render(<LogsPage />)
-    expect(screen.getByText('提醒设置（可选）')).toBeInTheDocument()
-  })
-
-  // handleAdd: hasReminder falsy path (branch 9)
-  it('renders reminder type select', () => {
-    render(<LogsPage />)
-    expect(screen.getByText('提醒类型')).toBeInTheDocument()
-  })
-
-  // Dialog onOpenChange: closes and resets (branch 18)
-  it('resets form when dialog closes', () => {
-    render(<LogsPage />)
-    expect(screen.getByText('添加案件日志')).toBeInTheDocument()
-  })
-
-  // search with caseName match (branch 12[2])
-  it('searches by case name in caseNameMap', () => {
-    mockQueryData = [makeLog({ case: 101, content: 'Log content' })]
-    render(<LogsPage />)
-    const input = screen.getByPlaceholderText('搜索日志内容、案件名称、操作人...')
-    fireEvent.change(input, { target: { value: 'nonexistent' } })
-    expect(screen.getByText('暂无日志')).toBeInTheDocument()
-  })
-
-  // Actor name fallback: real_name || username || '未知' (branch 117)
-  it('renders actor with real_name first', () => {
-    mockQueryData = [makeLog({ actor_detail: { real_name: 'Name', username: 'user' } })]
-    render(<LogsPage />)
-    expect(screen.getByText('Name')).toBeInTheDocument()
-  })
-
-  // grouped empty: created_at empty string (branch 129: empty dateKey -> skip)
-  it('skips logs with empty string created_at', () => {
-    mockQueryData = [
-      makeLog({ id: 1, content: 'Valid', created_at: '2025-06-01 10:00:00' }),
-      makeLog({ id: 2, content: 'Invalid', created_at: '' }),
-    ]
-    render(<LogsPage />)
-    expect(screen.getByText('Valid')).toBeInTheDocument()
-    expect(screen.queryByText('Invalid')).not.toBeInTheDocument()
-  })
-
-  // confirm button disabled with reminder type but no time (branch 19[3])
-  it('disables confirm button when reminder type set but no time', () => {
-    render(<LogsPage />)
-    const confirmBtn = screen.getByText('确认')
-    // Without caseId set, button is already disabled
-    expect(confirmBtn).toBeDisabled()
-  })
-
-  // Multiple logs on same date -> grid-cols-2 (line 260)
-  it('renders multi-column grid for multiple logs on same date', () => {
-    mockQueryData = [
-      makeLog({ id: 1, content: 'First', created_at: '2025-06-01 09:00:00' }),
-      makeLog({ id: 2, content: 'Second', created_at: '2025-06-01 14:00:00' }),
-    ]
-    render(<LogsPage />)
-    expect(screen.getByText('First')).toBeInTheDocument()
-    expect(screen.getByText('Second')).toBeInTheDocument()
-  })
-
-  // Single log on date -> no grid-cols-2 class (line 260)
-  it('renders single-column grid for single log on a date', () => {
-    mockQueryData = [
-      makeLog({ id: 1, content: 'Only', created_at: '2025-06-01 09:00:00' }),
-    ]
-    render(<LogsPage />)
-    expect(screen.getByText('Only')).toBeInTheDocument()
-  })
-
-  // Reminder type label fallback for unknown type (branch: getReminderLabel)
-  it('renders unknown reminder type as raw string', () => {
-    mockQueryData = [makeLog({
-      reminders: [{ id: 1, reminder_type: 'custom_type', due_at: '2025-06-15', is_completed: false }],
-    })]
-    render(<LogsPage />)
-    expect(screen.getByText('custom_type')).toBeInTheDocument()
-  })
-
-  // Reminder is_completed true (branch: r.is_completed &&)
-  it('shows check mark for completed reminder', () => {
-    mockQueryData = [makeLog({
-      reminders: [{ id: 1, reminder_type: 'hearing', due_at: '2025-06-15', is_completed: true }],
-    })]
-    render(<LogsPage />)
-    expect(screen.getByText('✓')).toBeInTheDocument()
-  })
-
-  // Reminder is_completed false
-  it('does not show check mark for uncompleted reminder', () => {
-    mockQueryData = [makeLog({
-      reminders: [{ id: 1, reminder_type: 'hearing', due_at: '2025-06-15', is_completed: false }],
-    })]
-    render(<LogsPage />)
-    expect(screen.queryByText('✓')).not.toBeInTheDocument()
-  })
-
-  // Reminder with null due_at
-  it('handles reminder with null due_at', () => {
-    mockQueryData = [makeLog({
-      reminders: [{ id: 1, reminder_type: 'hearing', due_at: null, is_completed: false }],
-    })]
-    render(<LogsPage />)
-    expect(screen.getByTestId('bell-icon')).toBeInTheDocument()
-  })
-
-  // Pagination: click load more (line 356-365)
-  it('shows load more when >10 groups and loads more on click', () => {
-    const logs = Array.from({ length: 15 }, (_, i) => makeLog({
-      id: i + 1, content: `Log ${i + 1}`,
-      created_at: `2025-06-${String(i + 1).padStart(2, '0')} 10:00:00`,
-    }))
-    mockQueryData = logs
-    render(<LogsPage />)
-    const btn = screen.queryByText(/加载更多/)
-    if (btn) {
-      fireEvent.click(btn)
-      expect(screen.getByText('Log 15')).toBeInTheDocument()
-    }
-  })
-
-  // Pagination: no load more when <10 groups
-  it('does not show load more when fewer than 10 groups', () => {
-    mockQueryData = [
-      makeLog({ id: 1, content: 'A', created_at: '2025-06-01 10:00:00' }),
-    ]
-    render(<LogsPage />)
-    expect(screen.queryByText(/加载更多/)).not.toBeInTheDocument()
-  })
-
-  // relativeDate boundary: exactly 7 days (branch 2[0] days < 7 is false)
-  it('shows date string for 7+ days ago', () => {
-    const old = new Date(Date.now() - 7 * 86400000)
-    const dateStr = `${old.getFullYear()}-${String(old.getMonth() + 1).padStart(2, '0')}-${String(old.getDate()).padStart(2, '0')} 10:00:00`
-    mockQueryData = [makeLog({ created_at: dateStr })]
-    render(<LogsPage />)
-    expect(screen.getByText('Test log')).toBeInTheDocument()
-  })
-})
