--- frontend/src/features/settings/__tests__/ServiceConfig.branches.test.tsx
@@ -1,345 +1,0 @@
-/**
- * Additional branch coverage tests for ServiceConfig.tsx
- * Focuses on uncovered branches not covered by the main test file.
- * Uses the SAME mock setup to avoid conflicts.
- */
-import { render, screen, fireEvent, waitFor, cleanup } from '@testing-library/react'
-import { useParams } from 'react-router'
-import { ServiceConfig } from '../components/ServiceConfig'
-import { useSystemConfigs } from '../hooks/use-system-configs'
-import { toast } from 'sonner'
-
-const mockUseParams = vi.mocked(useParams)
-
-vi.mock('sonner', () => ({
-  toast: {
-    success: vi.fn(),
-    error: vi.fn(),
-    info: vi.fn(),
-  },
-}))
-
-vi.mock('react-router', async (importOriginal) => {
-  const actual = await importOriginal<typeof import('react-router')>()
-  return { ...actual, useNavigate: () => vi.fn(), useParams: vi.fn().mockReturnValue({ category: 'llm' }) }
-})
-
-vi.mock('@/routes/paths', () => ({
-  PATHS: { ADMIN_SETTINGS: '/admin/settings' },
-}))
-
-vi.mock('@/lib/api', () => ({
-  getApiBaseUrl: () => 'http://localhost:8002/api/v1',
-  getBackendUrl: () => 'http://localhost:8002',
-  createFeatureApiClient: () => ({
-    get: () => ({ json: () => Promise.resolve([]) }),
-    post: () => ({ json: () => Promise.resolve({}) }),
-    put: () => ({ json: () => Promise.resolve({}) }),
-    patch: () => ({ json: () => Promise.resolve({}) }),
-    delete: () => ({ json: () => Promise.resolve({}) }),
-  }),
-}))
-
-const mockUpdateMutate = vi.fn()
-const mockCreateMutate = vi.fn()
-const mockPatchMutate = vi.fn()
-const mockDeleteMutate = vi.fn()
-
-vi.mock('../hooks/use-system-configs', () => ({
-  useSystemConfigs: vi.fn().mockReturnValue({
-    data: [{
-      category: 'llm',
-      items: [
-        { key: 'OPENAI_API_KEY', value: 'sk-test', category: 'llm', description: 'OpenAI API Key', is_secret: true, is_active: true, has_value: true },
-        { key: 'LLM_MODEL', value: 'gpt-4o', category: 'llm', description: 'Default model', is_secret: false, is_active: true, has_value: true },
-      ],
-    }],
-    isLoading: false,
-  }),
-  useUpdateSystemConfigs: () => ({
-    mutate: mockUpdateMutate,
-    isPending: false,
-  }),
-  useCreateSystemConfig: () => ({
-    mutate: mockCreateMutate,
-    isPending: false,
-  }),
-  usePatchSystemConfig: () => ({
-    mutate: mockPatchMutate,
-    isPending: false,
-  }),
-  useDeleteSystemConfig: () => ({
-    mutate: mockDeleteMutate,
-    isPending: false,
-  }),
-}))
-
-vi.mock('../constants/config-hints', () => ({
-  CATEGORY_HINTS: {
-    llm: {
-      title: 'LLM 配置',
-      description: '配置大语言模型相关参数',
-      fields: {
-        OPENAI_API_KEY: { label: 'OpenAI API Key', placeholder: 'sk-...', fullWidth: true },
-        LLM_MODEL: { label: '默认模型', placeholder: 'gpt-4o' },
-      },
-      fieldOrder: ['OPENAI_API_KEY', 'LLM_MODEL'],
-      groups: [],
-    },
-    grouped: {
-      title: '分组配置',
-      description: '测试分组',
-      fields: {
-        KEY_A: { label: 'Key A' },
-        KEY_B: { label: 'Key B' },
-        KEY_C: { label: 'Key C' },
-      },
-      fieldOrder: ['KEY_A', 'KEY_B', 'KEY_C'],
-      groups: [
-        { label: '组1', keys: ['KEY_A', 'KEY_B'] },
-        { label: '组2', keys: ['KEY_C'] },
-      ],
-    },
-    with_remaining: {
-      title: '带剩余',
-      description: '',
-      fields: { KEY_X: { label: 'Key X' }, KEY_Y: { label: 'Key Y' } },
-      fieldOrder: ['KEY_X', 'KEY_Y'],
-      groups: [{ label: '主要组', keys: ['KEY_X'] }],
-    },
-    system: { title: '系统连接', description: '' },
-    emptygroups: {
-      title: '空组',
-      description: '',
-      fields: {},
-      fieldOrder: [],
-      groups: [],
-    },
-    nohints: { title: '无提示', description: '无' },
-    secret_no_value: {
-      title: 'Secret No Value',
-      description: '',
-      fields: { MY_SECRET: { label: 'Secret Key' } },
-      fieldOrder: ['MY_SECRET'],
-      groups: [],
-    },
-  },
-}))
-
-import { useSystemConfigs } from '../hooks/use-system-configs'
-const mockUseSystemConfigs = useSystemConfigs as unknown as ReturnType<typeof vi.fn>
-const mockToast = vi.mocked(toast)
-
-describe('ServiceConfig - additional branch coverage', () => {
-  beforeEach(() => {
-    cleanup()
-    vi.clearAllMocks()
-    mockUseParams.mockReturnValue({ category: 'llm' })
-    mockUseSystemConfigs.mockReturnValue({
-      data: [{
-        category: 'llm',
-        items: [
-          { key: 'OPENAI_API_KEY', value: 'sk-test', category: 'llm', description: 'OpenAI API Key', is_secret: true, is_active: true, has_value: true },
-          { key: 'LLM_MODEL', value: 'gpt-4o', category: 'llm', description: 'Default model', is_secret: false, is_active: true, has_value: true },
-        ],
-      }],
-      isLoading: false,
-    })
-  })
-
-  // renderGroups: hintGroups.length > 0 (branch 127)
-  it('renders grouped fields with group labels', () => {
-    mockUseParams.mockReturnValue({ category: 'grouped' })
-    mockUseSystemConfigs.mockReturnValue({
-      data: [{
-        category: 'grouped',
-        items: [
-          { key: 'KEY_A', value: 'a', category: 'grouped', description: 'A', is_secret: false, is_active: true, has_value: true },
-          { key: 'KEY_B', value: 'b', category: 'grouped', description: 'B', is_secret: false, is_active: true, has_value: true },
-          { key: 'KEY_C', value: 'c', category: 'grouped', description: 'C', is_secret: false, is_active: true, has_value: true },
-        ],
-      }],
-      isLoading: false,
-    })
-    render(<ServiceConfig />)
-    expect(screen.getByText('组1')).toBeInTheDocument()
-    expect(screen.getByText('组2')).toBeInTheDocument()
-  })
-
-  // renderGroups: remaining fields (branch 144)
-  it('renders remaining fields with 其他 group label', () => {
-    mockUseParams.mockReturnValue({ category: 'with_remaining' })
-    mockUseSystemConfigs.mockReturnValue({
-      data: [{
-        category: 'with_remaining',
-        items: [
-          { key: 'KEY_X', value: 'x', category: 'with_remaining', description: 'X', is_secret: false, is_active: true, has_value: true },
-          { key: 'KEY_Y', value: 'y', category: 'with_remaining', description: 'Y', is_secret: false, is_active: true, has_value: true },
-        ],
-      }],
-      isLoading: false,
-    })
-    render(<ServiceConfig />)
-    expect(screen.getByText('主要组')).toBeInTheDocument()
-    expect(screen.getByText('其他')).toBeInTheDocument()
-  })
-
-  // secret field with has_value false (branch 420-429)
-  it('renders secret field with has_value false as 未设置', () => {
-    mockUseParams.mockReturnValue({ category: 'secret_no_value' })
-    mockUseSystemConfigs.mockReturnValue({
-      data: [{
-        category: 'secret_no_value',
-        items: [
-          { key: 'MY_SECRET', value: '', category: 'secret_no_value', description: 'Secret Key', is_secret: true, is_active: true, has_value: false },
-        ],
-      }],
-      isLoading: false,
-    })
-    render(<ServiceConfig />)
-    expect(screen.getByText('未设置')).toBeInTheDocument()
-  })
-
-  // empty groups: all empty fields (branch 371)
-  it('renders empty state when all groups empty', () => {
-    mockUseParams.mockReturnValue({ category: 'emptygroups' })
-    mockUseSystemConfigs.mockReturnValue({ data: [{ category: 'emptygroups', items: [] }], isLoading: false })
-    render(<ServiceConfig />)
-    expect(screen.getByText('该类别暂无配置项')).toBeInTheDocument()
-  })
-
-  // system category: test connection with empty URL (branch 159-163)
-  it('shows error when testing connection with empty backend URL', () => {
-    mockUseParams.mockReturnValue({ category: 'system' })
-    render(<ServiceConfig />)
-    const backendInput = screen.getByDisplayValue('http://localhost:8002')
-    fireEvent.change(backendInput, { target: { value: '' } })
-    fireEvent.click(screen.getByText('测试连通性'))
-    expect(screen.getByText('请先填写后端地址')).toBeInTheDocument()
-  })
-
-  // Edit dialog: toggle is_secret (branch 280)
-  it('handles edit with is_secret toggle change', () => {
-    render(<ServiceConfig />)
-    const editButtons = screen.getAllByTitle('编辑配置项')
-    fireEvent.click(editButtons[1])
-    // Change description to trigger modification
-    const descInput = screen.getByDisplayValue('Default model')
-    fireEvent.change(descInput, { target: { value: 'Updated' } })
-    const saveButtons = screen.getAllByText('保存')
-    fireEvent.click(saveButtons[saveButtons.length - 1])
-    expect(mockPatchMutate).toHaveBeenCalled()
-  })
-
-  // Edit: is_active toggle (branch 281)
-  it('handles edit with is_active toggle', () => {
-    render(<ServiceConfig />)
-    const editButtons = screen.getAllByTitle('编辑配置项')
-    fireEvent.click(editButtons[1])
-    expect(screen.getByText('启用')).toBeInTheDocument()
-    const descInput = screen.getByDisplayValue('Default model')
-    fireEvent.change(descInput, { target: { value: 'Updated' } })
-    const saveButtons = screen.getAllByText('保存')
-    fireEvent.click(saveButtons[saveButtons.length - 1])
-    expect(mockPatchMutate).toHaveBeenCalled()
-  })
-
-  // Edit: non-secret value change (branch 282)
-  it('handles edit with non-secret value change', () => {
-    render(<ServiceConfig />)
-    const editButtons = screen.getAllByTitle('编辑配置项')
-    fireEvent.click(editButtons[1])
-    // The edit dialog opens with current value. Change the value.
-    // Need to target the input in the edit dialog specifically
-    const descInput = screen.getByDisplayValue('Default model')
-    fireEvent.change(descInput, { target: { value: 'New description' } })
-    const saveButtons = screen.getAllByText('保存')
-    fireEvent.click(saveButtons[saveButtons.length - 1])
-    expect(mockPatchMutate).toHaveBeenCalled()
-  })
-
-  // Edit: secret value change (branch 283)
-  it('handles edit with secret value change', () => {
-    render(<ServiceConfig />)
-    const editButtons = screen.getAllByTitle('编辑配置项')
-    fireEvent.click(editButtons[0])
-    const valueInput = screen.getByPlaceholderText('留空则不修改')
-    fireEvent.change(valueInput, { target: { value: 'new-secret' } })
-    const saveButtons = screen.getAllByText('保存')
-    fireEvent.click(saveButtons[saveButtons.length - 1])
-    expect(mockPatchMutate).toHaveBeenCalled()
-  })
-
-  // Edit: no changes guard (branch 285)
-  it('shows info toast when editing with no changes', () => {
-    render(<ServiceConfig />)
-    const editButtons = screen.getAllByTitle('编辑配置项')
-    fireEvent.click(editButtons[1])
-    const saveButtons = screen.getAllByText('保存')
-    fireEvent.click(saveButtons[saveButtons.length - 1])
-    expect(mockToast.info).toHaveBeenCalledWith('没有需要保存的修改')
-  })
-
-  // No description for system category (branch 310-311)
-  it('does not show description for system category', () => {
-    mockUseParams.mockReturnValue({ category: 'system' })
-    render(<ServiceConfig />)
-    expect(screen.queryByText('配置大语言模型相关参数')).not.toBeInTheDocument()
-  })
-
-  // No backendItem: no edit/delete buttons (branch 399)
-  it('does not show edit/delete for system category', () => {
-    mockUseParams.mockReturnValue({ category: 'system' })
-    render(<ServiceConfig />)
-    expect(screen.queryAllByTitle('编辑配置项').length).toBe(0)
-    expect(screen.queryAllByTitle('删除配置项').length).toBe(0)
-  })
-
-  // fullWidth field rendering (branch 395)
-  it('renders fullWidth field correctly', () => {
-    render(<ServiceConfig />)
-    expect(screen.getByText('OpenAI API Key')).toBeInTheDocument()
-  })
-
-  // system category save with values (branch 225-231)
-  it('saves system config values to localStorage', () => {
-    mockUseParams.mockReturnValue({ category: 'system' })
-    render(<ServiceConfig />)
-    fireEvent.click(screen.getByText('保存配置'))
-    expect(localStorage.getItem('backend_url')).toBe('http://localhost:8002')
-    expect(mockToast.success).toHaveBeenCalledWith('系统连接配置已保存，刷新页面后生效')
-  })
-
-  // system category save with empty values (branch 228-230)
-  it('removes localStorage when system values empty', () => {
-    mockUseParams.mockReturnValue({ category: 'system' })
-    render(<ServiceConfig />)
-    const backendInput = screen.getByDisplayValue('http://localhost:8002')
-    fireEvent.change(backendInput, { target: { value: '' } })
-    const apiInput = screen.getByDisplayValue('http://localhost:8002/api/v1')
-    fireEvent.change(apiInput, { target: { value: '' } })
-    fireEvent.click(screen.getByText('保存配置'))
-    expect(localStorage.getItem('backend_url')).toBeNull()
-  })
-
-  // handleFieldChange for system category (branch 216-217)
-  it('handles system field change', () => {
-    mockUseParams.mockReturnValue({ category: 'system' })
-    render(<ServiceConfig />)
-    const input = screen.getByDisplayValue('http://localhost:8002')
-    fireEvent.change(input, { target: { value: 'http://new:9000' } })
-    expect(input).toHaveValue('http://new:9000')
-  })
-
-  // handleDelete guard (branch 300)
-  it('does not show delete dialog initially', () => {
-    render(<ServiceConfig />)
-    expect(screen.queryByText('确认删除配置项')).not.toBeInTheDocument()
-  })
-
-  // handleEdit guard (branch 277)
-  it('does not show edit dialog initially', () => {
-    render(<ServiceConfig />)
-    expect(screen.queryByText('编辑配置项')).not.toBeInTheDocument()
-  })
-})
